--- src/test.jsx
@@ -1,10 +1,0 @@
-import { createElement } from "@/utils/createElement";
-
-export default function Test() {
-  return (
-    <div className="test">
-      <span>1</span>
-      <span>2</span>
-    </div>
-  );
-}
